fixing decimal issue

+++ src/main/kotlin/com/example/runitup/mobile/model/User.kt
@@ -76,4 +76,4 @@ enum class UserType{
     ADMIN, SUPER_ADMIN
 }
 
-data class Coordinate(val longitude: Long, val latitude: Long)
+data class Coordinate(val longitude: Double, val latitude: Double)
